requirements.txt
Flask==2.3.2
flask-cors==3.1.0
tensorflow==2.20.0
numpy>=1.26,<2.0
scipy>=1.12,<2.0
gunicorn==21.2.0
Pillow==10.0.0